Query Creations
You have installed the employees database. We have reverse engineered the database to see the ER diagram. We have done preliminary data exploration. Do more in depth review of this database and come up with 5 important queries you can run - the queries cannot be just simple selects that does not give us any insight into the data. The queries you come with should provide valuable information about employees, the organization's salary structures and so on. The bottom line is - you have to think what information are important in making different decisions about the organization. Imagine yourself to the owner of the company or a manager in HR and so on. That should help. Upload a pdf file with the list of queries you could think of.
Make sure you state your name and the course name and the semester information in the submitted file.